+++ WOD/Timer/Detail/View/SubView/Setting/Timer/DetailTimerRestSet.swift
@@ -1,0 +1,44 @@
+//
+//  DetailTimerRestSet.swift
+//  Timer
+//
+//  Created by lkh on 1/1/24.
+//
+
+import SwiftUI
+
+// MARK: - SimpleTimerRestSet
+struct DetailTimerRestSet: View {
+    // MARK: Object
+    @EnvironmentObject var viewModel: DetailViewModel
+    @StateObject var manager = InputManager()
+    
+    // MARK: State/Binding
+    @State private var isChange: Bool = false
+    @Binding var showPopup: Bool
+    
+    // MARK: - View
+    var body: some View {
+        // MARK: main
+        NavigationView {
+            RestSet(selectedRestAmount: $manager.selectedRestAmount, isChange: $isChange,
+                    showPopup: $showPopup, isCalculatedBtn: $manager.isCalculatedBtn, manager: manager)
+        }
+        // MARK: side
+        .onAppear {
+            manager.selectedRestAmount = viewModel.selectedRestAmount
+        }
+        
+        .navigationTitle("Rest")
+        .navigationBarTitleDisplayMode(.inline)
+        
+        .popupSettingToolbar(
+            cancelAction:  { showPopup.toggle() },
+            action:  { isChange.toggle() },
+            completeAction: {
+                viewModel.selectedRestAmount = manager.selectedRestAmount
+                showPopup.toggle()
+            }
+        ) // popupSettingToolbar
+    } // body
+}
